feat: extract score route add info route
