+++ src/app/api/v1/group/info/route.ts
@@ -2,6 +2,7 @@ import { db } from "@/lib/services/firebase.admin";
 import { NextResponse } from "next/server";
 
 export async function GET() {
+  // Getall
   try {
     const snapshot = await db.collection("groups").get();
 
@@ -11,13 +12,13 @@ export async function GET() {
 
     const groups = snapshot.docs.map((doc) => ({
       id: doc.id,
-      score: doc.data()?.totalScore || 0,
+      ...doc.data(),
     }));
 
     return NextResponse.json({ groups }, { status: 200 });
   } catch (err) {
     return NextResponse.json(
-      { error: "Failed to get group scores: " + err },
+      { error: "Failed to get group information: " + err },
       { status: 500 }
     );
   }
